Fixed melee lifetap overflows - Because being healed for uint16s is not cool.

+++ zone/common.h
@@ -312,8 +312,8 @@ struct StatBonuses {
 	int16	FlurryChance;
 	int16	Accuracy[HIGHEST_SKILL+2];			//Accuracy/15 == % increase	[Spell Effect: Accuracy)
 	int16	HundredHands;						//extra haste, stacks with all other haste	i
-	int16	MeleeLifetap;						//i
-	int16	Vampirism;							//i
+	int32	MeleeLifetap;						//i
+	int32	Vampirism;							//i
 	int16	HealRate;							// Spell effect that influences effectiveness of heals
 	int32	MaxHPChange;						// Spell Effect
 	int16	SkillDmgTaken[HIGHEST_SKILL+2];		// All Skills + -1
